fix: update references to db column name
As per changes introduced in bug 39190, the
Koha::Files:Transport table is now file_transports
(which this plugin does not need to account for),
and had its unique identifier change from id to
file_transport_id.

Amend the code accordingly to prevent DB errors
and getting a 500 on save.

To replicate OR test:
1. navigate to /cgi-bin/koha/plugins/run.pl?class=
Koha::Plugin::AutomateMarcImport&method=configure
2. select a transport and profile
3. select save

Before this commit: 500
After: the new setting saves and displays as
expected